Add a hook to delay mounting until /media is writable.
If a device is inserted at boot time, it is possible for udev to start
uam before the root filesystem is mounted read/write. This leads to a
mount failure if none of the mountpoint locations exist.

Now uam supplies a pre-mount hook which delays the actual mounting until
${MOUNTPOINT_BASE} becomes writable, forking uam if necessary (to avoid
deadlocking udev).